requirements.txt
arabic-reshaper 
python-bidi
torch
pandas
seaborn
WordCloud